+++ descriptor/bridge_desc.h
@@ -56,15 +56,15 @@ public:
 	const char *get_name() const { return get_cursor()->get_name(); }
 	const char *get_copyright() const { return get_cursor()->get_copyright(); }
 
-	skin_desc_t const* get_cursor() const { return get_child<skin_desc_t>(2 + offset); }
+	skin_desc_t const* get_cursor() const { return get_child<skin_desc_t>((int)(2 + offset)); }
 
 	image_id get_background(img_t img, uint8 season) const {
 		const image_t *image = NULL;
 		if(season && number_of_seasons == 1) {
-			image = get_child<image_list_t>(3 + offset)->get_image(img);
+			image = get_child<image_list_t>((int)(3 + offset))->get_image(img);
 		}
 		if(image == NULL) {
-			image = get_child<image_list_t>(0 + offset)->get_image(img);
+			image = get_child<image_list_t>((int)(0 + offset))->get_image(img);
 		}
 		return image != NULL ? image->get_id() : IMG_EMPTY;
 	}
@@ -72,10 +72,10 @@ public:
 	image_id get_foreground(img_t img, uint8 season) const {
 		const image_t *image = NULL;
 		if(season && number_of_seasons == 1) {
-			image = get_child<image_list_t>(4 + offset)->get_image(img);
+			image = get_child<image_list_t>((int)(4 + offset))->get_image(img);
 		}
 		if(image == NULL) {
-			image = get_child<image_list_t>(1 + offset)->get_image(img);
+			image = get_child<image_list_t>((int)(1 + offset))->get_image(img);
 		}
 		return image != NULL ? image->get_id() : IMG_EMPTY;
 	}
